+++ tests/StreamTest.php
@@ -58,9 +58,11 @@ class StreamTest extends TestCase
     {
         set_error_handler(function () {
         });
-
+        error_clear_last();
+        $path = 'no/such/file.csv';
         $this->expectException(Exception::class);
-        Stream::createFromPath('no/such/file.csv');
+        $this->expectExceptionMessage('`'.$path.'`: failed to open stream: No such file or directory');
+        Stream::createFromPath($path);
 
         restore_error_handler();
     }
@@ -182,5 +184,21 @@ class StreamTest extends TestCase
         $this->assertSame($expected, $doc->getCsvControl());
         $this->expectException(Exception::class);
         $doc->setCsvControl(...['foo']);
+    }
+
+    /**
+     * @covers ::appendFilter
+     */
+    public function testAppendStreamFilterThrowsException()
+    {
+        set_error_handler(function () {
+        });
+        $filtername = 'foo.bar';
+        error_clear_last();
+        $this->expectException(Exception::class);
+        $this->expectExceptionMessage('Enable to locate filter `'.$filtername.'`');
+        $stream = Stream::createFromPath('php://temp', 'r+');
+        $stream->appendFilter($filtername, STREAM_FILTER_READ);
+        restore_error_handler();
     }
 }