+++ src/ByteSequence.php
@@ -4,14 +4,12 @@
 *
 * @license http://opensource.org/licenses/MIT
 * @link https://github.com/thephpleague/csv/
-* @version 9.1.0
+* @version 9.1.1
 * @package League.csv
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */
-declare(strict_types=1);
-
 namespace League\Csv;
 
 /**