+++ CHANGELOG.md
@@ -3,7 +3,6 @@
 All Notable changes to `Csv` will be documented in this file
 
 
-
 ## Next - TBD
 
 ### Added
@@ -17,6 +16,8 @@ All Notable changes to `Csv` will be documented in this file
 ### Fixed
 
 - issue with `error_get_last` usage when using a modified PHP error handler see [#254](https://github.com/thephpleague/csv/issues/254) - fixed by [@csiszarattila](https://github.com/csiszarattila)
+
+- Removed seekable word from Stream exception messages.
 
 ### Removed